refactor(ai): extract live-tool dispatcher from AssistantPanel
Move runAssistantLiveTool and all session_*, quick_command_*, screenshot,
and tutorial handlers (~480 lines) into assistantLiveTools.ts. Handlers
read live state from the workspace/dashboard stores and pane registry;
the panel-owned capabilities (navigation callbacks, tutorial overlay,
translations) are injected through an AssistantLiveToolDeps interface, so
the dispatcher is testable without mounting the panel. AssistantPanel.tsx
is now 2386 lines (3901 at session start).
